% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contincols.R
\name{contincols}
\alias{contincols}
\title{Create colours based on values}
\usage{
contincols(cols, colvals)
}
\arguments{
\item{cols}{A vector with two or more colours from which the colour scale is created by extrapolation between the colours}

\item{colvals}{A vector with values (for example values of a node attribute)}
}
\value{
A vector with colours that can be used directly to set the node colours of an \code{igraph} network plot.
}
\description{
\code{contincols()} creates a set of colours based on a set of values and any colour scale, such that the distance between the colours correspond to the distance between the values.
The colours can be used to colour the nodes of an \code{igraph} network plot, based on a node attribute.
}
\details{
For valid formats of cols see the col argument of the \code{col2rgb()} function.
Values in \code{colvals} must be between 0 and 1.
If the colours are used to colour the nodes of a network, then the length of \code{colvals} should equal the number of nodes in that network.
}
\examples{
# An example of colouring an igraph network based on node values:


# set colours for colour scale:
cols <- c('#CCFFFF','#003366')

# Make simulated node attribute values:
colvals <- c(rep(0.1,30), rep(0.4,30), rep(0.9,40))

# create colours for the network nodes, based on their attribute values:
nodecols <- contincols(cols, colvals)

# make a network and colour it:
library(igraph)
net <- sample_gnp(100,0.1)
V(net)$color <- nodecols
plot.igraph(net)
}
\seealso{
\code{\link{igraph}}
}
\author{
Josefine Bohr Brask, \email{bohrbrask@gmail.com}
}
