% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitnetbuild.R
\name{traitnetbuild}
\alias{traitnetbuild}
\title{Build a network matrix based on social attraction values}
\usage{
traitnetbuild(n, k, linktype, sociatvals)
}
\arguments{
\item{n}{An integer indicating the number of individuals (network nodes)}

\item{k}{An integer indicating the average degree (average number of links connected to each node)}

\item{linktype}{A string indicating the type of link ('stow', 'detw', or 'unw'; see Details)}

\item{sociatvals}{A vector with a social attraction value (a number between 0 and 1) for each pair of individuals (each dyad of nodes)}
}
\value{
An undirected network matrix (adjacency matrix). The matrix is symmetric, with zeros in the diagonal.
}
\description{
\code{traitnetbuild()} generates a network (an adjacency matrix) based on social attraction values,
such that pairs of individuals with higher social attraction are more likely to get a link.
}
\details{
The number of links is calculated based on the set average degree and number of individuals.
Then the dyads to be linked are found by a weighted draw, where dyads with higher social attraction have a higher chance of being picked.
The link weights are determined based on the set link type and added to the existing links.

To construct networks based on the trait preference model, it is recommended to use the \code{traitnet()} and \code{traitnetsmetrics()} functions.
Those functions call this one, and there is thus no need to use this one directly.

The number of individuals \code{n} must fit with the length of the \code{sociatvals} vector (which equals the number of dyads).

If there are not enough dyads with social attraction values above zero to give the input average degree,
then random links will be added among the remaining dyads to get the input average degree.

About \code{linktype}:
\itemize{
\item 'unw' gives unweighted (binary) links. Existing links are indicated with the value 1, non-existing links are indicated with 0.
\item 'stow' gives weighted links with stochasticity, where the weights are drawn from normal distributions with the social attraction values as means (sd = 0.05).
\item 'detw' gives deterministic weighted links, where the weight equals the social attraction of the dyad.
\item For weighted links, the link weights are in the interval [0,1], and existing links have a minimum weight of 0.001.
}
}
\examples{

n <- 100
k <- 10
linktype <- 'stow'

# For the example we use random values as social attraction values
# (see the examples in the `traitnetsociat()`  help page to generate
social attraction values based on the trait preference model):
ndyads <- (n*n-n)/2
sociatvals <-  runif(ndyads)

# Make the network matrix:
net <- traitnetbuild(n, k, linktype, sociatvals)
}
\references{
\url{https://arxiv.org/abs/2303.08107}
}
\seealso{
\code{\link{traitnet()}}
}
\author{
Josefine Bohr Brask, \email{bohrbrask@gmail.com}
}
