% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitnetsociat.R
\name{traitnetsociat}
\alias{traitnetsociat}
\title{Calculate social attraction values}
\usage{
traitnetsociat(traitclasses, preftypes, wvals, traitvals)
}
\arguments{
\item{traitclasses}{A vector of strings indicating the trait class for each trait ('cate' or 'cont'; see Details)}

\item{preftypes}{A vector of strings indicating the preference type to be used with each trait ('sim', or 'pop'; see Details)}

\item{wvals}{A vector with the importance (weight) of each trait-preference combination}

\item{traitvals}{A matrix with a column of trait values for each trait, and a row for each individual. Can be a vector if there is only a single trait}
}
\value{
A vector of social attraction values (one value for each pair of individuals).
}
\description{
\code{traitnetsociat()} calculates social attraction values between individuals, based on social preferences for traits (attributes).
The social attraction can be based on one or multiple traits, which may be used with different preference types.
}
\details{
The \code{traitnetsociat()} function calculates the social attraction for each pair of individuals, based on the two individuals'
trait values (attributes), and the preference type and importance (weight) set for each trait (such that traits with higher weight have more influence on the social attraction).
The calculation follows the preference model (see References).

To construct networks based on the trait preference model, it is recommended to use the \code{traitnet()} and \code{traitnetsmetrics()} functions.
Those functions call this one, and there is thus no need to use this one directly.

The function assumes that each input is ordered in the same way. For example, the first trait class given in the
\code{traitclasses} vector will be matched with the first preference type given in the \code{preftype} vector, the first
importance value given in the \code{wvals} vector, and the first column of trait values in the \code{traitvals} matrix. This also
means that the trait values for a given trait must be of the class given for that trait.

About \code{traitclasses}:
\itemize{
\item 'cate' indicates a categorical trait.
\item 'cont' indicates a continuous trait.
}

About \code{preftypes}:
\itemize{
\item There are two preference types: 'similarity' and 'popularity' (as described in Brask et al., see References).
\item 'sim' indicates that individuals prefer others that are similar to themselves in terms of the trait.
\item 'pop' indicates that individuals with high trait values are preferred (popular).
}

About \code{wvals}:
\itemize{
\item The sum of the \code{wval} vector, and each value in the vector, must lie in the interval [0,1], where 0 indicates no importance and 1 indicates full importance.
}

About \code{traitvals}:
\itemize{
\item The trait values must fit with the trait classes given in the \code{traitclasses} argument.
Categorical trait values should consist of integers that each indicates a category.
Continuous trait values should consist of numbers between 0 and 1. Empirical trait values on other scales should be transformed to this scale before being used.
}
}
\examples{

# An example of social attraction values that are based on the
# combined preferences for three traits of different type:

# Create trait values based on some trait types:
traittypes <- c('cate', 'tnorm', 'ranks')
n <- 100
tval1 <- traitvalues(n = n, traittype = traittypes[1], ncats = 2)
tval2 <- traitvalues(n = n, traittype = traittypes[2])
tval3 <- traitvalues(n = n, traittype = traittypes[3])
traitvals <- cbind(tval1, tval2, tval3)

# Set the class, preference type and importance for each trait
(The trait classes must fit with the trait values):
traitclasses <- c('cate', 'cont', 'cont')
preftypes <- c('sim', 'pop', 'sim')
wvals <- c(0.7, 0.1, 0.1)

# Generate the social attraction values:
sociatvals <- traitnetsociat(traitclasses, preftypes, wvals, traitvals)



}
\references{
\url{https://arxiv.org/abs/2303.08107}
}
\seealso{
\code{\link{traitnet()}}
}
\author{
Josefine Bohr Brask, \email{bohrbrask@gmail.com}
}
