% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitvalues.R
\name{traitvalues}
\alias{traitvalues}
\title{Generate trait values}
\usage{
traitvalues(n, traittype, ncats)
}
\arguments{
\item{n}{An integer indicating the number of trait values to be made (i.e. number of individuals for which the trait values are made)}

\item{traittype}{A string indicating the trait type ('cate', 'tnorm', or 'ranks'; see Details)}

\item{ncats}{An integer indicating the number of categories (only used if \code{traittype} is set to categorical)}
}
\value{
A vector of trait values. The vector has length \code{n}. The values are in random order. Non-categorical values are in the interval [0,1]; categorical values are given by integers.
}
\description{
\code{traitvalues()} generates a vector of trait values based on a given trait type.
}
\details{
To generate trait values, the function draws values from a distribution that corresponds to the given trait type.

To make networks based on trait preferences, it is recommended to use the \code{traitnet()} or \code{traitnetsmetrics()} functions.
Those functions call this function and there is no need to use this function directly.

About \code{traittype}:
\itemize{
\item 'cate'  gives categorical trait values. The values consist of the integers from 1 to \code{ncats}, which each indicates a category. Each category is drawn with the same probability.
\item 'tnorm' gives normally distributed trait values (truncated). The values are drawn from a truncated normal distribution that lie in the interval [0,1], with mean = 0.5 and standard deviation = 0.25.
\item 'ranks' gives trait values indicating ranks (e.g. social rank). The values are homogeneously distributed in the interval [0,1]. This corresponds to ranks indicated by integers from 1 to \code{n}, transformed to the interval [0,1].
}
}
\examples{
# Categorical trait values with two categories, for 100 individuals:
traitvalues(n = 100, traittype = 'cate', ncats = 2)

# Continuous trait values (from a truncated normal distribution), for 100 individuals:
traitvalues(n = 100, traittype = 'tnorm')

# Rank trait values, for 50 individuals:
traitvalues(n = 50, traittype = 'ranks')
}
\references{
\url{https://arxiv.org/abs/2303.08107}
}
\seealso{
\code{\link{traitnet()}}

\code{\link{traitnetsmetrics()}}
}
\author{
Josefine Bohr Brask, \email{bohrbrask@gmail.com}
}
